Add files via upload
One of my stronger assignments where I created classification trees to determine if a patient is likely to have heart disease based off a variety of medical factors. This first was completed by using "automatic" methods and then fine tuned to further increase accuracy. 

Next, this process was copied again but with a different dataset which is used to build models that predict if someone donated blood in the month of March based off their previous blood donation patterns. This was used using classification trees as well.

Both the R project and .rmd file is uploaded as well as the knitted word output file from the markdown.