Add files via upload
This is my final project for my predictive analytics class with UNCW. 
In the project I was working with a modified Ames Iowa dataset and was trying to predict which homes sold and will sell above the median home asking price at time of selling. 
In phase 1 conduct an EDA on the datasets and build a number of visualizations to help determine which variables may be a good predicator of the above median response variable.
In phase 2 I take the most promising variables and use them to build three model types: logistic regression, classification trees, and neural networks. Then comparing the pros and cons of each model.